refactor(error): extracted and renamed error module
- added doc comments for `error` module.
- renamed `NekosLifeError` to `Error`.
